+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.3] - 2020-6-xx
+
+- Add Typescript types to all components
+- Swap var(--font-secondary) with var(--font-primary)
+
 ## [1.10.2] - 2020-5-25
 
 - Add `CTALink` component